reafactor: Melhorar organização dos arquivos utilizando routers e controllers
Relacionado a #2
