+++ src/routes/index.ts
@@ -1,0 +1,6 @@
+import express from 'express';
+import { userRouter } from './userRouter';
+
+export const router = express.Router();
+
+router.use('/users', userRouter);
